backend/requirements.txt
# Web framework
fastapi>=0.95.0,<0.100.0
uvicorn>=0.22.0,<0.30.0
websockets>=11.0.0,<12.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.11.0,<2.0.0
psycopg2-binary>=2.9.6,<3.0.0  # PostgreSQL driver
python-dotenv>=1.0.0,<2.0.0    # Environment variable management

# Authentication and security
pyjwt>=2.6.0,<3.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0
bcrypt>=4.0.1,<5.0.0

# API client and networking
httpx>=0.24.0,<0.30.0
requests>=2.28.0,<3.0.0

# Data processing
pydantic>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0

# Monitoring and instrumentation
opentelemetry-api>=1.18.0,<2.0.0
opentelemetry-sdk>=1.18.0,<2.0.0
opentelemetry-exporter-otlp>=1.18.0,<2.0.0

# Resilience and fault tolerance
tenacity>=8.2.0,<9.0.0

# Utilities
python-dateutil>=2.8.2,<3.0.0
